+++ app/models/admin/research_article.rb
@@ -3,7 +3,15 @@ class Admin::ResearchArticle < ActiveRecord::Base
   # Constants
   # Attribute related macros
   # Associations
+  belongs_to :research_topic
+
   # Validations
+  validates :slug, presence: true
+  validates :title, presence: true
+  validates :description, presence: true
+  validates :content, presence: true
+  validates :author, presence: true
+
   # Callback
   # Other macros
   mount_uploader :photo, PhotoUploader