+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Added ability to flag industry sponsored clinical trials from admin interface
   - Added order option for clinical trials
   - Added interface to create new research articles
+    - Added ability to link approved research topic to research article
 
 ## 9.1.0 (December 7, 2015)